Added GUI Hide and show functions.
Added the ability to hide the main GUI to simulate a minimize to system tray function.  The script already had a system tray.  This required naming the main and start GUI so that the Hide/Show functions can properly target it.  Also, converted Xisumavoid's face, sourced from HermitCraft Fandom wiki (https://hermitcraft.fandom.com/wiki/XisumaVoid), into an ICO file for use by the script.  This is so the tray icon is unique from other AHK scripts to curb confusion for users who already have other scripts running.

Side note:  The overall GUI code can be cleaned up ALOT but using the named GUI method and giving the text controls vLabels for targeting.  This would allow for the GUI to be simply modified instead of being destroyed and rebuilt.  If you want me to do this I can.